+++ modules/video_player/manifest.py
@@ -24,7 +24,7 @@ MODULE = {
     },
 
     "description": "Video playback node for show media.",
-    "version": "1.3.7",
+    "version": "1.3.8",
     "apt_packages": [
         "vlc",
         "cec-utils"